Time: 0 ms (100.00%) | Memory: 9 MB (87.47%) - LeetSync

+++ 377-combination-sum-iv/combination-sum-iv.cpp
@@ -3,7 +3,7 @@ class Solution {
 public:
     int combinationSum4(vector<int>& nums, int target) {
         int n = nums.size();
-        vector<i128> dp(target + 1, 0);
+        vector<unsigned int> dp(target + 1, 0);
         dp[0] = 1;
 
         for (int j = 0; j <= target; j++) {
